requirements.txt
asgiref==3.3.1
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
idna==2.10
pycodestyle==2.6.0
python-decouple==3.4
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
